Avoid ambiguous exported data.
Fix issue #3 by checking if comma is the decimal separator, and if so,
use semi-colon instead to separate exported data.
I've tested that this doesn't change existing behaviour in the UK (which
uses '.' as the decimal separator).  I've not tested yet in a locale
where comma is the decimal separator.