Fix issue #6 by giving Arduino more time after reset.
Recent changes mean that we now fail with a warning if
singleTact_.Initialise fails.  It seems to fail if we write to the
Arduino's serial port too soon after opening it.

This patch doubles our 'settling time' to two seconds, which fixes the
problem as tested on my PC.  Five tests in this configuration all
succeeded.

I think the initialise call has always failed but the old code just
never tested it, and possibly failed in some other way.